+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Adding `#.forEachNeighborUntil`.
 * Adding `#.updateNode`.
 * Adding `#.updateEdge`.
+* Adding `attributes` to `nodeAttributesUpdated` events.
 * Optimizing `#.edges(A, B)`.
 * Optimizing `#.forEachEdge(A, B)`.
 * Optimizing `#.mergeEdge`.